move all service to a dedicated directory
