+++ internal/api/front/perses/src/app/project/dashboard/service/dashboard.service.ts
@@ -13,11 +13,11 @@
 
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { ErrorHandlingService } from '../../shared/service/error-handling.service';
+import { ErrorHandlingService } from '../../../shared/service/error-handling.service';
 import { Observable } from 'rxjs';
-import { UrlBuilderUtil } from '../../shared/utils/url-builder.util';
+import { UrlBuilderUtil } from '../../../shared/utils/url-builder.util';
 import { catchError } from 'rxjs/operators';
-import { DashboardModel } from './dashboard.model';
+import { DashboardModel } from '../dashboard.model';
 
 @Injectable({
   providedIn: 'root'
