+++ internal/api/front/perses/src/app/project/dashboard/service/event-feed.service.ts
